finished example

+++ example/order/src/main/java/cz/filipklimes/diploma/framework/example/order/business/ShoppingCartItem.java
@@ -2,6 +2,8 @@ package cz.filipklimes.diploma.framework.example.order.business;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 public class ShoppingCartItem
 {
 
@@ -9,9 +11,9 @@ public class ShoppingCartItem
     private final Integer productId;
 
     @Getter
-    private final Integer quantity;
+    private final BigDecimal quantity;
 
-    public ShoppingCartItem(final Integer productId, final Integer quantity)
+    public ShoppingCartItem(final Integer productId, final BigDecimal quantity)
     {
         this.productId = productId;
         this.quantity = quantity;
